+++ src/cards/entity-card/entity-card-editor.ts
@@ -12,6 +12,13 @@ import { loadHaComponents } from "../../utils/loader";
 import { ENTITY_CARD_EDITOR_NAME } from "./const";
 import { EntityCardConfig, entityCardConfigStruct } from "./entity-card-config";
 
+const ENTITY_LABELS = [
+  "timer_enabled",
+  "timer_duration",
+  "motion_enabled",
+  "motion_sensor",
+];
+
 const SCHEMA: HaFormSchema[] = [
   { name: "entity", selector: { entity: {} } },
   { name: "name", selector: { text: {} } },
@@ -28,6 +35,35 @@ const SCHEMA: HaFormSchema[] = [
     ],
   },
   ...APPEARANCE_FORM_SCHEMA,
+  {
+    type: "grid",
+    name: "",
+    schema: [
+      { name: "timer_enabled", selector: { boolean: {} } },
+      {
+        name: "timer_duration",
+        selector: { number: { min: 1, max: 3000, step: 1, unit_of_measurement: "seconds" } },
+      },
+    ],
+  },
+  {
+    type: "grid",
+    name: "",
+    schema: [
+      { name: "motion_enabled", selector: { boolean: {} } },
+      {
+        name: "motion_sensor",
+        selector: { 
+          entity: { 
+            domain: ["binary_sensor"],
+            filter: {
+              device_class: "motion"
+            }
+          } 
+        },
+      },
+    ],
+  },
   ...computeActionsFormSchema(),
 ];
 
@@ -53,6 +89,9 @@ export class EntityCardEditor
 
     if (GENERIC_LABELS.includes(schema.name)) {
       return customLocalize(`editor.card.generic.${schema.name}`);
+    }
+    if (ENTITY_LABELS.includes(schema.name)) {
+      return customLocalize(`editor.card.entity.${schema.name}`);
     }
     return this.hass!.localize(
       `ui.panel.lovelace.editor.card.generic.${schema.name}`